Day 8 2023
part2t.py was how I started once I realized brute force took too long. Figuring out that each ending node for the paths is on a repeating step value of N*p where p is a prime times the number of instructions, the rest is easy to compute.
I'm assuming part2.py works for valid inputs from AoC, but it will not work for any random input. part2t.py would need to be completed for that.